+++ src/App.js
@@ -3,6 +3,7 @@ import './App.css';
 import DronePilot from './Pages/DronePilot'; 
 import Home from './Pages/Home';
 import Header from './Components/Header';
+import Request from './Pages/Request';
 import Explora from './Pages/Explora';
 import UserRegistry from './Pages/UserRegistry';
 import Error404 from './Pages/Error404';
@@ -18,6 +19,7 @@ function App() {
           <Route exact path="/explora" component={Explora} />
           <Route exact path="/pilot-profile" component={DronePilot} />
           <Route exact path="/user-registry" component={UserRegistry} />
+          <Route exact path="/request" component={Request} />
           <Route exact path="/login" component={Login} />
           <Route exact path="*" component={Error404}/>
         </Switch>
