+++ src/Pages/Request.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import Querie from '../Components/Querie'
+import GlobalContainer from '../utils/GlobalStyles'
+
+
+function Request(){
+
+return (
+  <GlobalContainer>
+    <Querie/>
+  </GlobalContainer>
+  )
+}
+
+export default Request
